requirements-dev.txt
asgiref==3.3.1
backcall==0.2.0
decorator==4.4.2
dj-database-url==0.5.0
Django==3.1.7
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
ipdb==0.13.5
ipython==7.21.0
ipython-genutils==0.2.0
jedi==0.18.0
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.16
ptyprocess==0.7.0
Pygments==2.8.0
PyJWT==2.0.1
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
traitlets==5.0.5
wcwidth==0.2.5

